requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in requirements.in
#
aiohttp==3.7.4.post0
    # via discord-py
async-timeout==3.0.1
    # via aiohttp
attrs==22.1.0
    # via
    #   aiohttp
    #   pytest
black==22.6.0
    # via -r requirements-dev.in
build==0.8.0
    # via pip-tools
chardet==4.0.0
    # via aiohttp
click==8.1.3
    # via
    #   black
    #   pip-tools
discord-py==1.7.3
    # via -r requirements.in
flake8==5.0.4
    # via -r requirements-dev.in
idna==3.3
    # via yarl
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements-dev.in
loguru==0.6.0
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy==0.971
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   pytest
pathspec==0.9.0
    # via black
pep517==0.13.0
    # via build
pip-tools==6.8.0
    # via -r requirements-dev.in
platformdirs==2.5.2
    # via black
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pycodestyle==2.9.1
    # via flake8
pyflakes==2.5.0
    # via flake8
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via -r requirements-dev.in
python-dotenv==0.20.0
    # via -r requirements.in
tomli==2.0.1
    # via
    #   black
    #   build
    #   mypy
    #   pep517
    #   pytest
typing-extensions==4.3.0
    # via
    #   aiohttp
    #   mypy
wheel==0.37.1
    # via pip-tools
yarl==1.8.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
